Move handle-move handler to itemtool
Now all should work without GTK
